Have been unable to utilize values from the dropdown box and check box
Have been attmepting to pull values from the elements in order to use
them in if statements but have not had success.  Also have attempted to
change the color of the input fields only when the error class has been
applied to make it easier to read without success.

Have not been able to get the reset function to clear any messages
displayed from previous attmepts to enter information,  Both the success
and error messages stay until overridden with new messages from the
submit function.